The bug with condition is fixed
The problem was that the System.currentTime() does not work as I
expected. Maybe the cycle, which manipulates the data, runs very fast
and every time all the conditions get the same ID. But this problem
cannot be seen in debugging because there is time to be generated
different "value".